+++ src/components/FilterBarInline/filterBarInlineChangelogData.ts
@@ -10,6 +10,15 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "2024-08-29",
+    version: "3.3.0",
+    type: "Update",
+    affects: ["Documentation", "Functionality"],
+    notes: [
+      "Adds closeOnBlur prop to the render prop function to apply closeOnBlur to MultiSelect children.",
+    ],
+  },
+  {
     date: "2024-07-25",
     version: "3.2.0",
     type: "New Feature",