+++ src/components/Select/selectChangelogData.ts
@@ -10,6 +10,15 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "2024-08-29",
+    version: "3.3.0",
+    type: "Update",
+    affects: ["Functionality"],
+    notes: [
+      "Updates component to accept a `defaultValue` prop to set initial value of uncontrolled components.",
+    ],
+  },
+  {
     date: "2024-06-20",
     version: "3.1.6",
     type: "Update",